Move linode unit tests to match module layout.
